pip_requirements.txt
Django==1.4.1
Pillow==1.7.7
South==0.7.5
django-piston==0.2.3
django-annoying==0.7.6
django-guardian==1.0.4
django-userena==1.1.2
easy-thumbnails==1.0.3
